backend/requirements.txt
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Cors==6.0.1
PyMySQL==1.1.2
python-dotenv==1.1.1
alembic==1.17.0
Werkzeug==3.1.3
cryptography==46.0.2
PyJWT==2.10.1